+++ packages/core/src/memory/index.ts
@@ -1,2 +1,3 @@
 export * from './types';
 export * from './memory';
+export * from './mock';
